requirements.txt
requests==2.32.4
python-dotenv==1.1.0
gradio==5.33.1
pypdf==5.6.0
pypdf2==3.0.1
openai==1.85.0
openai-agents==0.0.17
ipywidgets==8.1.7
sendgrid==6.12.3
huggingface_hub

